requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.5
aiosignal==1.3.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
decorator==4.4.2
discord.py==2.4.0
frozenlist==1.5.0
idna==3.10
imageio==2.36.0
imageio-ffmpeg==0.5.1
moviepy==1.0.3
multidict==6.1.0
numpy==2.1.3
pillow==11.0.0
proglog==0.1.10
propcache==0.2.0
pytubefix==8.5.1
requests==2.32.3
tqdm==4.67.0
urllib3==2.2.3
yarl==1.17.2


